// apps/api/src/constants/historyActionType.ts
export enum HistoryActionType {
  // User actions
  USER_CREATED = 'USER_CREATED',
  USER_DEACTIVATED = 'USER_DEACTIVATED',
  USER_REACTIVATED = 'USER_REACTIVATED',
  USER_UPDATED = 'USER_UPDATED',
  USER_ROLE_CHANGED = 'USER_ROLE_CHANGED',
  
  // Project actions
  PROJECT_CREATED = 'PROJECT_CREATED',
  PROJECT_SUPERVISOR_ASSIGNED = 'PROJECT_SUPERVISOR_ASSIGNED',
  PROJECT_SUPERVISOR_CHANGED = 'PROJECT_SUPERVISOR_CHANGED',
  PROJECT_EMPLOYEE_ADDED = 'PROJECT_EMPLOYEE_ADDED',
  PROJECT_EMPLOYEE_REMOVED = 'PROJECT_EMPLOYEE_REMOVED',
  PROJECT_DELETED = 'PROJECT_DELETED',
  PROJECT_RESTORED = 'PROJECT_RESTORED',
  
  // Team actions
  TEAM_CREATED = 'TEAM_CREATED',
  TEAM_SUPERVISOR_ASSIGNED = 'TEAM_SUPERVISOR_ASSIGNED',
  TEAM_SUPERVISOR_CHANGED = 'TEAM_SUPERVISOR_CHANGED',
  TEAM_MEMBER_ADDED = 'TEAM_MEMBER_ADDED',
  TEAM_MEMBER_REMOVED = 'TEAM_MEMBER_REMOVED',
  TEAM_DELETED = 'TEAM_DELETED',
}